# openapi.yaml
openapi: 3.0.3
info:
  title: CORECROW API
  version: 1.0.0
  description: Black Polar shared backend. Cookie identity, tenant membership
    authorization, audited contract commerce. /api aliases are deprecated.
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Secure-better-auth.session_token
      description: Better Auth cookie; development uses better-auth.session_token.
...
paths:
...
  /v1/users:
    get:
      summary: List global users (operator only, maximum 50)
      tags:
        - Identity
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      minLength: 1
                      maxLength: 128
                    name:
                      type: string
                      nullable: true
                    email:
                      type: string
                      format: email
                    role:
                      type: string
                      enum:
                        - USER
                        - DEVELOPER
                        - ADMIN
                        - SUPERADMIN
                    emailVerified:
                      type: boolean
                    createdAt:
                      type: string
                      format: date-time
                  required:
                    - id
                    - name
                    - email
                    - role
                    - emailVerified
                    - createdAt
                  additionalProperties: false
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      requestId:
                        type: string
                    required:
                      - code
                      - message
                      - requestId
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      requestId:
                        type: string
                    required:
                      - code
                      - message
                      - requestId
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
    post:
      summary: Create a credential identity (operator only)
      tags:
        - Identity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 12
                  maxLength: 128
                role:
                  type: string
                  enum:
                    - USER
                    - DEVELOPER
                    - ADMIN
                    - SUPERADMIN
                  default: USER
              required:
                - name
                - email
                - password
              additionalProperties: false
      security:
        - sessionCookie: []
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    minLength: 1
                    maxLength: 128
                  name:
                    type: string
                    nullable: true
                  email:
                    type: string
                    format: email
                  role:
                    type: string
                    enum:
                      - USER
                      - DEVELOPER
                      - ADMIN
                      - SUPERADMIN
                  emailVerified:
                    type: boolean
                  createdAt:
                    type: string
                    format: date-time
                required:
                  - id
                  - name
                  - email
                  - role
                  - emailVerified
                  - createdAt
                additionalProperties: false
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      requestId:
                        type: string
                    required:
                      - code
                      - message
                      - requestId
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      requestId:
                        type: string
                    required:
                      - code
                      - message
                      - requestId
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
  /v1/users/{id}:
    patch:
      summary: Update a profile; global role changes require superadmin
      tags:
        - Identity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                role:
                  type: string
                  enum:
                    - USER
                    - DEVELOPER
                    - ADMIN
                    - SUPERADMIN
              additionalProperties: false
      parameters:
        - schema:
            type: string
            minLength: 1
            maxLength: 128
          in: path
          name: id
          required: true
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    minLength: 1
                    maxLength: 128
                  name:
                    type: string
                    nullable: true
                  email:
                    type: string
                    format: email
                  role:
                    type: string
                    enum:
                      - USER
                      - DEVELOPER
                      - ADMIN
                      - SUPERADMIN
                  emailVerified:
                    type: boolean
                  createdAt:
                    type: string
                    format: date-time
                required:
                  - id
                  - name
                  - email
                  - role
                  - emailVerified
                  - createdAt
                additionalProperties: false
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      requestId:
                        type: string
                    required:
                      - code
                      - message
                      - requestId
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      requestId:
                        type: string
                    required:
                      - code
                      - message
                      - requestId
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
    delete:
      summary: Delete an identity without memberships (superadmin only)
      tags:
        - Identity
      parameters:
        - schema:
            type: string
            minLength: 1
            maxLength: 128
          in: path
          name: id
          required: true
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                enum:
                  - "null"
                nullable: true
        4XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      requestId:
                        type: string
                    required:
                      - code
                      - message
                      - requestId
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
        5XX:
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                      requestId:
                        type: string
                    required:
                      - code
                      - message
                      - requestId
                    additionalProperties: false
                required:
                  - error
                additionalProperties: false
...
  /v1/identity/legacy-migration:
    post:
      summary: Opt-in legacy credential enrollment; revokes old sessions and
        permanently retires the legacy ID
      tags:
        - Identity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 254
                password:
                  type: string
                  minLength: 12
                  maxLength: 128
              required:
                - email
                - password
              additionalProperties: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                required:
                  - migrated
                properties:
                  migrated:
                    enum:
                      - true
...
  /api/users/{id}:
    get:
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
    patch:
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
    delete:
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
...
  /api/admin/keys/{id}:
    delete:
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
...
servers:
  - url: https://api.blackpolar.org